BAU: Refactoring and cleanup
