+++ integration-tests/jest.config-mocked.ts
@@ -6,6 +6,6 @@ import type { Config } from "@jest/types";
 import commonConfig from "./jest.config.common";
 const config: Config.InitialOptions = {
   ...commonConfig,
-  testMatch: ["**/tests/mocked/*.test.ts"],
+  testMatch: ["**/tests/mocked/**/*.test.ts"],
 };
 export default config;
