--- integration-tests/tests/jest.config.ts
@@ -1,12 +1,0 @@
-/*
- * For a detailed explanation regarding each configuration property and type check, visit:
- * https://jestjs.io/docs/configuration
- */
-import type { Config } from "@jest/types";
-import commonConfig from "./jest.config.common";
-
-const config: Config.InitialOptions = {
-  ...commonConfig,
-  testMatch: ["**/tests/**/*.test.ts"],
-};
-export default config;
